Add netstandard2.1 support to Utf8String.Experimental.
