+++ src/libraries/System.Private.CoreLib/src/System.Private.CoreLib.Shared.projitems
@@ -1817,7 +1817,6 @@
     <Compile Include="$(MSBuildThisFileDirectory)System\Utf8String.Conversion.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)System\Utf8String.Enumeration.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)System\Utf8String.Manipulation.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)System\Utf8String.netcoreapp.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)System\Utf8String.Searching.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)System\Utf8StringSplitOptions.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)System\Text\Utf8Span.cs" />
@@ -1825,7 +1824,6 @@
     <Compile Include="$(MSBuildThisFileDirectory)System\Text\Utf8Span.Conversion.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)System\Text\Utf8Span.Enumeration.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)System\Text\Utf8Span.Manipulation.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)System\Text\Utf8Span.netcoreapp.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)System\Text\Utf8Span.Searching.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)System\Text\Utf8StringComparer.cs" />
   </ItemGroup>
