--- src/libraries/System.Private.CoreLib/src/System/Text/Utf8Span.netcoreapp.cs
@@ -1,19 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-using System.Buffers;
-using System.ComponentModel;
-using System.Diagnostics;
-using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
-using System.Text.Unicode;
-using Internal.Runtime.CompilerServices;
-
-namespace System.Text
-{
-    public readonly ref partial struct Utf8Span
-    {
-        // TODO: eerhardt delete me
-    }
-}
