+++ src/libraries/System.Utf8String.Experimental/src/System/IO/Utf8StringStream.cs
@@ -58,7 +58,7 @@ namespace System.IO
         }
 
         public
-#if NETCOREAPP
+#if !NETSTANDARD2_0
             override
 #endif
             int Read(Span<byte> buffer)
@@ -80,7 +80,7 @@ namespace System.IO
             return Task.FromResult(Read(new Span<byte>(buffer, offset, count)));
         }
 
-#if NETCOREAPP
+#if !NETSTANDARD2_0
         public override ValueTask<int> ReadAsync(Memory<byte> buffer, CancellationToken cancellationToken = default)
         {
             return new ValueTask<int>(Read(buffer.Span));
@@ -128,13 +128,13 @@ namespace System.IO
 
         public override void Write(byte[] buffer, int offset, int count) => throw new NotSupportedException();
 
-#if NETCOREAPP
+#if !NETSTANDARD2_0
         public override void Write(ReadOnlySpan<byte> buffer) => throw new NotSupportedException();
 #endif
 
         public override Task WriteAsync(byte[] buffer, int offset, int count, CancellationToken cancellationToken) => throw new NotSupportedException();
 
-#if NETCOREAPP
+#if !NETSTANDARD2_0
         public override ValueTask WriteAsync(ReadOnlyMemory<byte> buffer, CancellationToken cancellationToken = default) => throw new NotSupportedException();
 #endif
         public override void WriteByte(byte value) => throw new NotSupportedException();
