--- src/libraries/System.Utf8String.Experimental/src/System/Net/Http/Utf8StringContent.netcoreapp3.cs
@@ -1,19 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-using System.Buffers;
-using System.IO;
-using System.Net.Http.Headers;
-using System.Runtime.InteropServices;
-using System.Threading;
-using System.Threading.Tasks;
-
-namespace System.Net.Http
-{
-    public sealed partial class Utf8StringContent : HttpContent
-    {
-        protected override Task SerializeToStreamAsync(Stream stream, TransportContext? context) =>
-            stream.WriteAsync(_content.AsMemoryBytes()).AsTask();
-    }
-}
