--- src/libraries/System.Utf8String.Experimental/src/System/Net/Http/Utf8StringContent.netcoreapp5.cs
@@ -1,22 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-using System.Buffers;
-using System.IO;
-using System.Net.Http.Headers;
-using System.Runtime.InteropServices;
-using System.Threading;
-using System.Threading.Tasks;
-
-namespace System.Net.Http
-{
-    public sealed partial class Utf8StringContent : HttpContent
-    {
-        protected override Task SerializeToStreamAsync(Stream stream, TransportContext? context) =>
-            SerializeToStreamAsync(stream, context, default);
-
-        protected override Task SerializeToStreamAsync(Stream stream, TransportContext? context, CancellationToken cancellationToken) =>
-            stream.WriteAsync(_content.AsMemoryBytes(), cancellationToken).AsTask();
-    }
-}
